Cleanup, verify weights
