+++ programs/castle-vault/src/backend_container/rate.rs
@@ -1,6 +1,20 @@
-use solana_maths::Rate;
+use anchor_lang::prelude::ProgramError;
+use boolinator::Boolinator;
+use solana_maths::{Rate, TryAdd};
+
+use crate::errors::ErrorCode;
 
 use super::BackendContainer;
+
+impl BackendContainer<Rate> {
+    pub fn verify_weights(&self) -> Result<(), ProgramError> {
+        let sum = self
+            .into_iter()
+            .map(|(_p, r)| r)
+            .try_fold(Rate::zero(), |acc, x| acc.try_add(*x))?;
+        (sum != Rate::one()).as_result((), ErrorCode::StrategyError.into())
+    }
+}
 
 impl From<BackendContainer<u16>> for BackendContainer<Rate> {
     fn from(c: BackendContainer<u16>) -> Self {
