+++ programs/castle-vault/src/instructions/rebalance.rs
@@ -43,7 +43,7 @@ impl From<&Allocations> for RebalanceDataEvent {
 pub enum Reserves {
     Solend(SolendReserve),
     Port(PortReserve),
-    Jet(jet::state::Reserve),
+    Jet(Box<jet::state::Reserve>),
 }
 
 impl AccountDeserialize for Reserves {
@@ -60,12 +60,8 @@ impl AccountDeserialize for Reserves {
 impl Pack for Reserves {
     const LEN: usize = 42;
 
-    fn pack_into_slice(&self, dst: &mut [u8]) {
-        match self {
-            Reserves::Solend(v) => v.pack_into_slice(dst),
-            Reserves::Port(v) => v.pack_into_slice(dst),
-            Reserves::Jet(_v) => todo!(), //v.pack_into_slice(dst),
-        }
+    fn pack_into_slice(&self, _dst: &mut [u8]) {
+        todo!()
     }
 
     fn unpack_from_slice(_src: &[u8]) -> Result<Self, ProgramError> {
@@ -220,8 +216,11 @@ pub fn handler_chris(
                 proposed_weights
             );
 
+            proposed_weights.verify_weights()?;
+
             let proposed_apr = get_apr_chris(&proposed_weights, &proposed_allocations, assets)?;
             let proof_apr = get_apr_chris(&strategy_weights, &strategy_allocations, assets)?;
+
             #[cfg(feature = "debug")]
             msg!(
                 "Proposed APR: {:?}\nProof APR: {:?}",
